backport of commit 650e783b060fea8d96e4f84ba8e1169df2372de7
